Call callback() if set to do some post-processing (#70)
* Call callback() if set to do some post-processing

Treat it as a proof of concept.
I want to keep the script output, but add some tweaks, e.g. wrap img with a div to show the image alt data on hover.
This PR allows me to call callback() after the script output. When options.get_data is set to true, callback() works as before.

* Update jquery.instagramFeed.js

* Update jquery.instagramFeed.js